<Feat> : domain 추가

+++ hamonize-center/src/main/java/com/controller/curl/CurlAgentBackupController.java
@@ -47,7 +47,7 @@ public class CurlAgentBackupController {
 		pcUuid = pcUuid.trim();
 
 		// uuid로 부서정보 가져오기
-		int segSeq = pcUUID(pcUuid);
+		Long segSeq = pcUUID(pcUuid);
 		if (segSeq == 0) {
 			return "nodata";
 		}
@@ -136,11 +136,11 @@ public class CurlAgentBackupController {
 	 * 
 	 * @return 부서seq
 	 */
-	public int pcUUID(String pcUUID) {
+	public Long pcUUID(String pcUUID) {
 		GetAgentJobVo agentVo = new GetAgentJobVo();
 		agentVo.setPc_uuid(pcUUID);
 		agentVo = agentJobMapper.getAgentJobPcUUID(agentVo);
-		int segSeq = 0;
+		Long segSeq = 0L;
 		if (agentVo != null) {
 			segSeq = agentVo.getSeq();
 		}
