+++ hamonize-center/src/main/java/com/controller/curl/CurlAgentRecoveryController.java
@@ -40,7 +40,7 @@ public class CurlAgentRecoveryController {
 		String output = "";
 		uuid = uuid.trim();
 
-		int segSeq = pcUUID(uuid);
+		Long segSeq = pcUUID(uuid);
 		if (segSeq == 0) {
 			return "nodata";
 		}
@@ -105,12 +105,12 @@ public class CurlAgentRecoveryController {
 		return jsonObject;
 	}
 
-	public int pcUUID(String uuid) {
+	public Long pcUUID(String uuid) {
 		GetAgentJobVo agentVo = new GetAgentJobVo();
 		agentVo.setPc_uuid(uuid);
 		agentVo = agentJobMapper.getAgentJobPcUUID(agentVo);
 
-		int segSeq = 0;
+		Long segSeq = 0L;
 		if (agentVo != null) {
 			segSeq = agentVo.getSeq();
 		}
