+++ hamonize-center/src/main/java/com/mapper/IUserMapper.java
@@ -26,7 +26,7 @@ public interface IUserMapper {
 
 	public OrgVo getUserOrgPath(int seq);
 
-	public OrgVo getUserNewOrgPath(int seq);
+	public OrgVo getUserNewOrgPath(Long seq);
 
 	public UserVo getUserInfo(int seq);
